// src/app/interfaces/paciente.interface.ts
export interface Paciente {
  _id:                          string;
  abrev_tipo_doc:               string;
  num_doc:                      string;
  paciente:                     string;
  desc_genero:                  string;
  desc_etn:                     string;
  fecha_nacimiento:             string;
  id_ubigeo_res:                string;
  departamento_residencia:      string;
  provincia_residencia:         string;
  distrito_residencia:          string;
  actividad:                    string;
  anioedad_atencion:            string;
  mesedad_atencion:             string;
  diaedad_atencion:             string;
  fecha_vacunacion:             string;
  id_vac_pac:                   string;
  codigo_vacuna:                string;
  vacuna:                       string;
  dosis_aplicada:               string;
  gruporiesgo:                  string;
  vacunador:                    string;
  DISA:                         string;
  RED:                          string;
  MICRORED:                     string;
  cod_unico:                    string;
  establecimiento:              string;
  departamento_establecimiento: string;
  provincia_establecimiento:    string;
  distrito_establecimiento:     string;
  fecha_registro:               string;
  registrador:                  string;
  id_estrategiavac:             string;
  id_tipoactividad:             string;
  Lote_Vacuna:                  string;
  Fabricante:                   string;
  Fabricante_Abrev:             string;
  NOMBRE_CENTRO_VACUNACION:     string;
}